+++ internal/rules/rule.go
@@ -18,7 +18,8 @@ type Rule interface {
 
 func All() []Rule {
 	return []Rule{
-		SkewRule{}, GCRule{}, SpillRule{}, FailedTasksRule{}, TinyTasksRule{}, IdleStageRule{}, ExecutorSupplyRule{},
+		SkewRule{}, GCRule{}, SpillRule{}, SchedulerDelayRule{}, RemoteShuffleRule{}, SpeculativeTasksRule{},
+		FailedTasksRule{}, TinyTasksRule{}, IdleStageRule{}, ExecutorSupplyRule{},
 	}
 }